refactor(tests): split schema parity and SNMP client test modules
